chore: lower linecheck gate to 1000 lines
Split four oversized test files (cli_tests, http_tests,
routine_storage_tests, service_tests) into sibling modules so every
.rs file stays under 1000 lines. Updates .githooks/pre-push to
enforce the new limit.
