+++ .changeset/linecheck-1000.md
@@ -1,0 +1,5 @@
+---
+"moadim": patch
+---
+
+chore: lower linecheck gate from 1500 to 1000 lines
