+++ src/routes/http.rs
@@ -528,3 +528,7 @@ pub async fn run_with_listener_until(
 #[cfg(test)]
 #[path = "http_tests.rs"]
 mod http_tests;
+
+#[cfg(test)]
+#[path = "http_listener_tests.rs"]
+mod http_listener_tests;
